validation/requirements.txt
Flask
rich
gunicorn
requests
flask-cors
boto3
python-dotenv